refactor: rename type files, add a TODO note
