+++ frontend/app/pages/event/[eventId]/book/seats/(components)/CartSummary.vue
@@ -1,7 +1,7 @@
 <script setup lang="ts">
 import { ref } from 'vue'
-import type { Ticket } from '../(types)/ticket.type'
-import type { CartItem } from '../(types)/event-payment.type'
+import type { Ticket } from '../(types)/ticket'
+import type { CartItem } from '../(types)/eventPayment'
 
 const { t: translate, locale } = useI18n()
