+++ frontend/app/pages/event/[eventId]/book/seats/(types)/event.ts
@@ -1,3 +1,4 @@
+// TODO: Resolve no content error
 export interface Event {
   id:    string
   title: string
