+++ frontend/app/pages/event/[eventId]/book/seats/(types)/seatingLayout.ts
@@ -1,3 +1,4 @@
+// TODO: Resolve no content error
 export interface SeatPos {
   x: number
   y: number
