+++ frontend/app/pages/event/[eventId]/book/seats/(types)/ticket.ts
@@ -1,3 +1,4 @@
+// TODO: Resolve no content error
 export interface Ticket {
   id:             string
   name:           string
